requirements.txt
numpy
opencv-python
pillow
torch
torchvision
pytesseract
pyyaml
scikit-learn
tqdm
